Fixes 2738: support SCRAM-SHA-512 sasl mech for kafka
